North American HIV Data Analysis
In this project, I collaborated with three peers to build a project from the ground up. We collected our data from open source, went through the cleaning, organized our dataset with SQL, and finished it off with Tableau Public for the visualizations. I have attached the Tableau Public file with a clear report on the North American HIV Data Analysis from 2007 to 2020.